+++ FeralBright/Config.cs
@@ -6,29 +6,28 @@ public static class Config
 {
     public static class Flashlight
     {
-        public static readonly FloatConfig Intensity = new FloatConfig()
-            .WithDefinition("Flashlight", "Intensity")
+        public static readonly FloatConfig Intensity = new FloatConfig("Flashlight", "Intensity")
             .WithDescription("Determines the intensity of the light. The higher the value, the brighter the light will be.")
             .WithDefaultValue(800F)
-            .WithRange(0F, 25_000F)
+            .WithMin(0F)
+            .WithMax(25_000F)
             .WithStep(1F);
 
-        public static readonly FloatConfig Range = new FloatConfig()
-            .WithDefinition("Flashlight", "Range")
+        public static readonly FloatConfig Range = new FloatConfig("Flashlight", "Range")
             .WithDescription("Determines the range of the light. The higher the value, the further the light will reach.")
             .WithDefaultValue(99_999F)
-            .WithRange(0F, 99_999F)
+            .WithMin(0F)
+            .WithMax(99_999F)
             .WithStep(1F);
 
-        public static readonly FloatConfig Spread = new FloatConfig()
-            .WithDefinition("Flashlight", "Spread")
+        public static readonly FloatConfig Spread = new FloatConfig("Flashlight", "Spread")
             .WithDescription("Determines the spread of the light. The higher the value, the wider the light will be.")
             .WithDefaultValue(130F)
-            .WithRange(1F, 179F)
+            .WithMin(1F)
+            .WithMax(179F)
             .WithStep(1F);
 
-        public static readonly ColorConfig Color = new ColorConfig()
-            .WithDefinition("Flashlight", "Color")
+        public static readonly ColorConfig Color = new ColorConfig("Flashlight", "Color")
             .WithDescription(
                 """
                 Determines the color of the light. The color is represented as a hexadecimal string.
@@ -39,8 +38,7 @@ public static class Config
             )
             .WithDefaultValue("#FFFFFF");
 
-        public static readonly FloatConfig OffsetX = new FloatConfig()
-            .WithDefinition("Flashlight", "X Offset")
+        public static readonly FloatConfig OffsetX = new FloatConfig("Flashlight", "X Offset")
             .WithDescription(
                 """
                 Determines the offset of the light on the X-axis. A value of 1 will place the light 1 unit to the right of the player.
@@ -48,11 +46,11 @@ public static class Config
                 """
             )
             .WithDefaultValue(0F)
-            .WithRange(-10F, 10F)
+            .WithMin(-10F)
+            .WithMax(10F)
             .WithStep(0.1F);
 
-        public static readonly FloatConfig OffsetY = new FloatConfig()
-            .WithDefinition("Flashlight", "Y Offset")
+        public static readonly FloatConfig OffsetY = new FloatConfig("Flashlight", "Y Offset")
             .WithDescription(
                 """
                 Determines the offset of the light on the Y-axis. A value of 0 will place the light at the player's eye level.
@@ -60,11 +58,11 @@ public static class Config
                 """
             )
             .WithDefaultValue(1.5F)
-            .WithRange(-10F, 10F)
+            .WithMin(-10F)
+            .WithMax(10F)
             .WithStep(0.1F);
 
-        public static readonly FloatConfig OffsetZ = new FloatConfig()
-            .WithDefinition("Flashlight", "Z Offset")
+        public static readonly FloatConfig OffsetZ = new FloatConfig("Flashlight", "Z Offset")
             .WithDescription(
                 """
                 Determines the offset of the light on the Z-axis. A value of 1 will place the light 1 unit in front of the player.
@@ -72,11 +70,11 @@ public static class Config
                 """
             )
             .WithDefaultValue(0F)
-            .WithRange(-10F, 10F)
+            .WithMin(-10F)
+            .WithMax(10F)
             .WithStep(0.1F);
 
-        public static readonly FloatConfig Pitch = new FloatConfig()
-            .WithDefinition("Flashlight", "Pitch")
+        public static readonly FloatConfig Pitch = new FloatConfig("Flashlight", "Pitch")
             .WithDescription(
                 """
                 Determines the pitch of the light. The pitch is the side-to-side rotation of the light.
@@ -84,11 +82,11 @@ public static class Config
                 """
             )
             .WithDefaultValue(0F)
-            .WithRange(-180F, 180F)
+            .WithMin(-180F)
+            .WithMax(180F)
             .WithStep(1F);
 
-        public static readonly FloatConfig Yaw = new FloatConfig()
-            .WithDefinition("Flashlight", "Yaw")
+        public static readonly FloatConfig Yaw = new FloatConfig("Flashlight", "Yaw")
             .WithDescription(
                 """
                 Determines the yaw of the light. The yaw is the up-and-down rotation of the light.
@@ -96,15 +94,14 @@ public static class Config
                 """
             )
             .WithDefaultValue(0F)
-            .WithRange(-180F, 180F)
+            .WithMin(-180F)
+            .WithMax(180F)
             .WithStep(1F);
     }
 
     public static class Sun
     {
-
-        public static readonly FloatConfig InsideIntensity = new FloatConfig()
-            .WithDefinition("Sun", "Inside Intensity")
+        public static readonly FloatConfig InsideIntensity = new FloatConfig("Sun", "Inside Intensity")
             .WithDescription(
                 """
                 Determines the intensity of the sun's light while inside the facility.
@@ -112,11 +109,11 @@ public static class Config
                 """
             )
             .WithDefaultValue(4F)
-            .WithRange(0F, 10F)
+            .WithMin(0F)
+            .WithMax(10F)
             .WithStep(1F);
 
-        public static readonly BoolConfig EnablePositionOverride = new BoolConfig()
-            .WithDefinition("Sun", "Enable Position Override")
+        public static readonly BoolConfig EnablePositionOverride = new BoolConfig("Sun", "Enable Position Override")
             .WithDescription(
                 """
                 Determines whether the sun's position should be overridden.
@@ -125,8 +122,7 @@ public static class Config
             )
             .WithDefaultValue(true);
 
-        public static readonly FloatConfig PositionOverride = new FloatConfig()
-            .WithDefinition("Sun", "Position Override")
+        public static readonly FloatConfig PositionOverride = new FloatConfig("Sun", "Position Override")
             .WithDescription(
                 """
                 Determines the lifecycle position of the sun. The higher the value, the closer the sun will be to the end of its lifecycle.
@@ -136,7 +132,8 @@ public static class Config
                 """
             )
             .WithDefaultValue(0.333F)
-            .WithRange(0F, 0.99F)
+            .WithMin(0F)
+            .WithMax(1F)
             .WithStep(0.001F);
     }
 }